Tratamento erros handling

+++ src/main/java/com/devsuperior/dscommerce/services/ProductService.java
@@ -1,31 +1,30 @@
 package com.devsuperior.dscommerce.services;
 
-import java.util.List;
-import java.util.Optional;
-
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.devsuperior.dscommerce.dto.ProductDTO;
 import com.devsuperior.dscommerce.entities.Product;
+import com.devsuperior.dscommerce.exceptions.DatabaseException;
+import com.devsuperior.dscommerce.exceptions.ResourceNotFoundException;
 import com.devsuperior.dscommerce.repositories.ProductRepository;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
 	@Autowired
 	private ProductRepository repository;
+	
 	@Transactional(readOnly = true)
-	public ProductDTO findById(Long id) {
-		Optional<Product> result = repository.findById(id); //busca no banco o produto = id
-		Product product = result.get();// atribui o resutl set para a variavel product
-		//agora tem que copiar os dados do result set(result) para o produto product. Para não fazer isso abaixo tem a criação no contrutor
-		//do ProductDTO
-		//ProductDTO dto = new ProductDTO(product.getId(), product.getName(), product.getDescription()......
-		ProductDTO dto = new ProductDTO(product); // copia para a variavel dto o product para fornecer ao controlador
-		return dto;		
+	public ProductDTO findById(Long id) {		
+		Product product = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado")) ; 
+		return new ProductDTO(product); 
 	}
 	
 	@Transactional(readOnly = true)
@@ -53,24 +52,35 @@ public class ProductService {
 
 	@Transactional
 	public ProductDTO update(Long id, ProductDTO dto) {
-		//diferente do método findbyid. Não vai no banco e faz sql. Só prepara 
-		Product entity = repository.getReferenceById(id);
-		//copiar para a entidade os dados que vieram da dto(la do controler)
-		/*entity.setName(dto.getName());
-		entity.setDescription(dto.getDescription());
-		entity.setPrice(dto.getPrice());
-		entity.setimgUrl(dto.getImgUrl());
-		entity = repository.save(entity);
-		*/
-		copyDTOToEntity(dto, entity);
-		
-		return new ProductDTO(entity);		
-		
+		try{//diferente do método findbyid. Não vai no banco e faz sql. Só prepara 
+			Product entity = repository.getReferenceById(id);
+			//copiar para a entidade os dados que vieram da dto(la do controler)
+			/*entity.setName(dto.getName());
+			entity.setDescription(dto.getDescription());
+			entity.setPrice(dto.getPrice());
+			entity.setimgUrl(dto.getImgUrl());
+			entity = repository.save(entity);
+			*/
+			copyDTOToEntity(dto, entity);
+			
+			return new ProductDTO(entity);		
+		}
+		catch (EntityNotFoundException e){
+			throw new ResourceNotFoundException("Recurso não encontrado."); 
+		}
 	}
 	
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		repository.deleteById(id);
+		if (!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		try {
+	        	repository.deleteById(id);    		
+		}
+	    	catch (DataIntegrityViolationException e) {
+	        	throw new DatabaseException("Falha de integridade referencial");
+	   	}
 	}
 
 	private void copyDTOToEntity(ProductDTO dto, Product entity) {
